make to OGR conversions (not to *and* from OGR conversions) streamable by default

+++ hoot-core/src/main/cpp/hoot/core/io/OsmXmlWriter.h
@@ -136,8 +136,8 @@ private:
   int _encodingErrorCount;
   std::shared_ptr<QXmlStreamWriter> _writer;
   geos::geom::Envelope _bounds;
-  long _numWritten;
-  long _statusUpdateInterval;
+  int _numWritten;
+  int _statusUpdateInterval;
   AddExportTagsVisitor _addExportTagsVisitor;
 
   static QString _typeName(ElementType e);
